Fix 42 non-matching functions in pool-b files
- 10 mov_w_vs_x: change function pointer cast types from s64 to s32
  for zero-valued arguments (mov w1,wzr vs mov x1,xzr)
- 10 ldr_vs_add: add missing vtable indirection (*(s64*)(...))
  to load function pointers instead of using address as pointer
- 19 and_vs_mov: change extern decls from u32 to u64 return type
  and add & 0xFFFFFFFF mask for proper zero-extension encoding
- 3 prologue_swap: attempted asm("") barriers (unfixable without
  enhancing fix_prologue.py for non-zero SP offsets)

Verified: 243 → 285 (+42), 28.4% of pool-b functions

+++ src/app/fun_batch_d_016.cpp
@@ -22,7 +22,7 @@ extern s64 DAT_710532e7c8;
 [[noreturn]] void FUN_7103344c38(void) { FUN_71039c20a0(); }
 
 // 0x71033625b8 — vtable call at *param_1+8, return 0 (56 bytes)
-u64 FUN_71033625b8(s64 *param_1)
+u32 FUN_71033625b8(s64 *param_1)
 {
     (*(void(*)())(*(s64*)(*param_1 + 8)))();
     return 0;
